Add cli module
